requirements.txt
kafka-python
pyspark
psycopg2-binary
apache-airflow
